requirement.txt
# LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Vector store and text processing
faiss-cpu>=1.7.4
numpy>=1.24.0

# Environment and HTTP requests
python-dotenv>=1.0.0
requests>=2.31.0

# Type annotations (for Python < 3.10)
typing-extensions>=4.8.0

# Optional: For better performance with FAISS
# faiss-gpu>=1.7.4  # Uncomment if you have CUDA GPU

# Optional: For visualization (if you want to use graph visualization)
# graphviz>=0.20.0
# pygraphviz>=1.11